Today, we worked on adjusting the pre-set powers for the shooter as well as pre-set positions for the hood in order to be able to score on the goal from three different ranges.
At the same time, we worked on fixing the gamepads' loop since there were subsystems that were interfering with one another.

You might notice that we commented a turret's configuration in order to use its encoder besides the camera in order to program a preset angle from the farthest shooting range.

Keep in mind that for now, we're living the camera disabled since we need to spend more time adjusting the factor of the Tx value.